Исправлены ошибки указанные в комментариях к пулл-реквесту
