+++ TickPic/Structures/PSStruct.swift
@@ -1,0 +1,36 @@
+//
+//  PSStruct.swift
+//  TickPic
+//
+//  Created by Кирилл Марьясов on 25.05.2024.
+//
+
+import Foundation
+
+struct ProfileResult: Codable {
+    let username: String
+    let firstName: String
+    let lastName: String?
+    let bio: String?
+    
+    enum CodingKeys: String, CodingKey {
+        case username
+        case firstName = "first_name"
+        case lastName = "last_name"
+        case bio
+    }
+}
+
+struct Profile {
+    let username: String
+    let name: String
+    let loginName: String
+    let bio: String
+    
+    init(username: String, firstName: String, lastName: String?, bio: String?) {
+        self.username = username
+        self.name = "\(firstName) \(lastName ?? "")"
+        self.loginName = "@\(username)"
+        self.bio = bio ?? ""
+    }
+}
