+++ TickPic/Utilities/Constants.swift
@@ -11,7 +11,9 @@ enum Constants {
     static let secretKey = "RtcVfP5lt6DNpTlSsFBglUeXNxIAGIwMhx34NVxPCVs"
     static let redirectURI = "urn:ietf:wg:oauth:2.0:oob"
     static let accessScope = "public+read_user+write_likes"
+    
     static let defaultBaseURL = URL(string: "https://api.unsplash.com")!
     static let unsplashAuthorizeURLString = "https://unsplash.com/oauth/authorize"
     static let unsplashTokenURLString = "https://unsplash.com/oauth/token"
+    static let meEndpoint = "\(defaultBaseURL)/me"
 }
